+++ server/react/src/HomePage/HomePage.js
@@ -1,33 +1,29 @@
-import React from "react";
+import React, { useState } from "react";
 import { Link } from "react-router-dom";
 import { connect } from "react-redux";
 
-class HomePage extends React.Component {
+function HomePage({ user }) {
+  const [images, setImages] = useState([]);
 
-  render() {
-    const { user, users } = this.props;
-    return (
-      <div className="col-md-6 col-md-offset-3">
-        <h1>Hi {user.firstName}!</h1>
-        <p>You're logged in with React & JWT!!</p>
-        <h3>Users from secure api end point:</h3>
-        {users.loading && <em>Loading users...</em>}
-        {users.error && (
-          <span className="text-danger">ERROR: {users.error}</span>
-        )}
-        {users.items && (
-          <ul>
-            {users.items.map((user, index) => (
-              <li key={user.id}>{user.firstName + " " + user.lastName}</li>
-            ))}
-          </ul>
-        )}
-        <p>
-          <Link to="/login">Logout</Link>
-        </p>
-      </div>
+  const onDrop = acceptedFiles => {
+    setImages(
+      acceptedFiles.map(file =>
+        Object.assign(file, {
+          preview: URL.createObjectURL(file)
+        })
+      )
     );
-  }
+  };
+
+  return (
+    <div className="col-md-6 col-md-offset-3">
+      <h1>Hello, {user.username}!</h1>
+      <p>You're logged in with React & JWT!!</p>
+      <p>
+        <Link to="/login">Logout</Link>
+      </p>
+    </div>
+  );
 }
 
 function mapStateToProps(state) {
