+++ server/resources.py
@@ -46,6 +46,7 @@ class UserLogin(Resource):
             refresh_token = create_refresh_token(identity=data['username'])
             return {
                 'user': {
+                    'username': data['username'],
                     'message': 'Logged in as {}'.format(current_user.username),
                     'access_token': access_token,
                     'refresh_token': refresh_token
